Factor out environment operation in a new module
