+++ lib/local-cluster/lib/Cardano/Wallet/Launch/Cluster.hs
@@ -56,9 +56,7 @@ import Cardano.Wallet.Launch.Cluster.Cluster
     )
 import Cardano.Wallet.Launch.Cluster.ClusterEra
     ( ClusterEra (..)
-    , clusterEraFromEnv
     , clusterEraToString
-    , localClusterConfigsFromEnv
     )
 import Cardano.Wallet.Launch.Cluster.ClusterM
     ( ClusterM
@@ -68,6 +66,10 @@ import Cardano.Wallet.Launch.Cluster.Config
     ( Config (..)
     , ShelleyGenesisModifier
     , TestnetMagic (..)
+    )
+import Cardano.Wallet.Launch.Cluster.Env
+    ( clusterEraFromEnv
+    , localClusterConfigsFromEnv
     )
 import Cardano.Wallet.Launch.Cluster.Faucet
     ( sendFaucetAssetsTo
